requirements.txt
appdirs==1.4.3
Django==1.10.6
django-debug-toolbar==1.7
djangorestframework==3.6.2
factory-boy==2.8.1
Faker==0.7.10
inflection==0.3.1
packaging==16.8
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
pytest-django==3.1.2
pytest-factoryboy==1.3.1
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
sqlparse==0.2.3
